Update _staged.py
Minor bug fix. When setting hyperstage=initial_staging, there was an error because there were no new stagings to test. So I added a condition in the while loop that the number of new stagings must be non-zero. If there are no new models to test (when hyperstage=initial_staging), the model selection step is now skipped and the staging from initial_staging is returned.